Get rid of Babel at run-time
