+++ package.json
@@ -4,9 +4,6 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
-    "babel-preset-env": "^1.4.0",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.24.1",
     "express": "^4.15.2",
     "ws": "^2.3.1"
   },
