+++ server/index.js
@@ -1,4 +1,81 @@
-require('babel-register')({
-  presets: [['env', { targets: { node: 'current' } }], 'stage-0']
+const { request, createServer } = require('http');
+
+const WebSocket = require('ws');
+const Server = WebSocket.Server;
+const express = require('express');
+
+const port = process.env.PORT || 8080;
+
+const baseOptions = {
+  method: 'GET',
+  socketPath: '/var/run/docker.sock',
+};
+
+const dockerAPIRequest = path => {
+  return new Promise((res, rej) => {
+    let buffer = '';
+
+    const r = request({ ...baseOptions, path }, response => {
+      response.on('data', chunk => (buffer = buffer + chunk));
+      response.on('end', () => res(buffer));
+    });
+
+    r.on('error', rej);
+
+    r.end();
+  });
+};
+
+const fetchData = () =>
+  Promise.all([
+    dockerAPIRequest('/nodes').then(JSON.parse),
+    dockerAPIRequest('/services').then(JSON.parse),
+    dockerAPIRequest('/networks').then(JSON.parse),
+    dockerAPIRequest('/tasks').then(JSON.parse),
+  ]).then(([nodes, services, networks, tasks]) => ({
+    nodes,
+    services,
+    networks,
+    tasks,
+  }));
+
+// start the polling
+
+let latestData = {};
+setInterval(() => {
+  fetchData().then(it => (latestData = it));
+}, 500);
+
+// set up the application
+
+const app = express();
+
+app.use(express.static('client'));
+app.get('/_health', (req, res) => res.end());
+app.get('/data', (req, res) => {
+  fetchData().then(it => res.send(it)).catch(e => res.send(e.toString()));
 });
-require('./server');
+
+// set up the server
+
+const server = createServer();
+const wsServer = new Server({
+  path: '/stream',
+  server,
+});
+
+server.on('request', app);
+
+wsServer.on('connection', ws => {
+  const interval = setInterval(() => {
+    if (ws.readyState === WebSocket.OPEN) {
+      ws.send(JSON.stringify(latestData, null, 2));
+    }
+  }, 200);
+
+  ws.on('close', () => clearInterval(interval));
+});
+
+server.listen(port, () => {
+  console.log(`Listening on ${port}`);
+});
